feat(httpapi): annotate Descriptor with RequestType (#1010)
This diff changes the httpapi to annotate a Descriptor with the
RequestType. Internally, the descriptor will include a RequestDescriptor
that for now only contains the body. Soon, it may contain more fields.

With the Descriptor annotated with a request type, we're now well
positioned to generate half of the swagger.

The reference issue is https://github.com/ooni/probe/issues/2372