feat(httpapi): annotate Descriptor with ResponseType
This diff modifies httpapi to annotate the descriptor with the
ResponseType, which also allows us to write code where we naturally
call an API as follows:

```Go
resp, err := httpapi.Call(ctx, desc, epnt)
```

and where `resp` is of the correct response type and where `desc`
has been configured to unmarshal using the JSON format.

In addition to this syntactic-sugar change, this diff matters because
knowing the type of the request and of the response allows us to
generate a swagger to compare to the server's swagger.

Part of https://github.com/ooni/probe/issues/2372.