refactor(session): observe full check-in response (#1037)
This diff refactors engine/session.go such that the session code can observe the full check-in response. We're still returning the nettests subset to the caller, since normally callers only care about that. By observing the full response, we'll be able to cache the feature flags, which will then be useful to implement https://github.com/ooni/probe/issues/2396.