fix(httpapi): don't return nil, nil with null JSON input (#1038)
While working on https://github.com/ooni/probe/issues/2396, I noticed
that the httpapi code could return nil, nil when the input from the
server was the `null` UTF-8 string. In such a case, golang translates
the `null` string into a `nil` pointer. However, if we pass the
pointer to a struct rather than the pointer to a pointer, the code
returns an empty struct, which is much safer.

This reasoning works as long as we assume the caller wants to get
a struct by pointer or nil on error. If the caller for some reason
wants the code to return a `**T` type, we're back in the case where
the code can return `nil`, `nil`. But all our code always wants
a `*T` back and this seems the only case that makes sense.