feat: mechanism compile and embed tor (#1052)
This diff introduces a mechanism to compile and embed tor. We plan on using this functionality on Android and iOS but we have not finished working on updating the build yet. In the meanwhile, here's an interim result. We're committing the code to run unit and integration testing of this new functionality on linux/amd64.

The reference issue is https://github.com/ooni/probe/issues/2365. 

The Go based build was sketched out in https://github.com/ooni/probe/issues/2401.