fix: attempt to make all workflows green
The check for the number of tokens was failing for Windows so
relax the check to say "three or more".

There's a warning related to use the old checkout@v2, so upgrade
to using checkout@v3 for all workflows in the repo.

The Linux action didn't set the correct version of Go, so it
wasn't possible to start the build.

Everything else was WAI.

Part of https://github.com/ooni/probe/issues/2273.