fix(netxlite): allow overriding default cert pool
This diff tweaks https://github.com/ooni/probe-cli/pull/1068
to make sure overriding the default cert pool works.

In https://github.com/ooni/probe-cli/pull/1068 we introduced
code to add this functionality but we never tested it was working
as intended. It turns out it was not!

Because this diff amends the previous diff, we'll consider it
part of https://github.com/ooni/probe/issues/2135.