chore: don't build libtorlinux on every pull request
We have been running this workflow very often at the beginning to
see if it was completing successfully. Now that we have a bit more
confidence in that, let's make it run less often.

Specifically, I chose to let it run as often as the alltests.yml
workflow, which seems frequently enough.