[backport] fix(android): sync compiler flags with ndk-build (#1071)
It seems using sanitizers leads to builds that SEGFAULT
as soon as we attempt to use the AAR generated by go mobile.

So, let's sync up with the flags used by ndk-build, which
have still some hardening but don't use sanitizers.

While there, run the libtorlinux.yml workflow less frequently
now that we have some confidence it's WAI.

Also, fix the MONOREPO workflow to produce an Android build,
which was using an old makefile target.

Closes https://github.com/ooni/probe/issues/2404