[backport] fix(libtor): don't leak a file descriptor (#1072)
Discovered while trying to understand the reason why we have an abort caused by Android's fdsan.

This is not the reason why we abort, but it seems correct to fix this issue anyway.

While there, notice that stopping to run libtorlinux for every PR (which we just did in https://github.com/ooni/probe-cli/commit/5ebcca256c50e7b40d3141534627d151ab20e4bd) was wrong. We need to run this workflow for every PR because it runs tests for the libtor integration.

Part of https://github.com/ooni/probe/issues/2405