fix: use the embedded control conn
It turns out previously we were not using the control conn. Using
such a conn is not the default and needs to be configured.

However, the code at https://github.com/ooni/probe-cli/pull/1052
tries to be immutable and hence has a different algorithm than the
one inside go-libtor. In particular, this new code always creates
a control connection and then closes it. So, definitely the new
code I wrote is such that it triggers the tor issue.

On the other end, to hope to see the issue with go-libtor we must
enable using the control conn. Let us do that and see.

Part of https://github.com/ooni/probe/issues/2405.