fix(netxlite): use private static default cert pool
This diff is slightly more complex than the naive diff that creates
a private static pool and uses it for quic and tls. The reason is
that we also need to support a mechanism allowing integration tests
to transparently override the default cert pool.

See https://github.com/ooni/probe/issues/2413 for context.