refactor(oohelperd): improve handler tests (#1101)
This diff attempts to rewrite and steer handler_test.go such that
it mostly contains unit tests. As part of this work, we create a
single factory function for constructing the handler. While there,
we also rename the indexLogger to prefixLogger (a more accurate name)
and we deploy more comments and whitespaces.

Part of https://github.com/ooni/probe/issues/2413