chore(webconnectivitylte): bump to 0.5.21
We don't need to backport 8b96164c8c7ad26d58a57f41aae955b8dc3a6967 to
the release/3.17 branch, because such a diff does not make sense in the
context of this branch. Still, it makes sense to bump the version
number of webconnectivity LTE to mean that we applied a compatible
change to what we applied in the master branch (and such a change
happens to be doing nothing in this branch).