fix: only construct cert pool when needed (#1129)
This pull request avoids constructing a cert pool when we can use the default used by netxlite. We also rename the `NewDefaultCertPool` to `NewMozillaCertPool` for clarity. Additionally, renaming ensures that code that was using the previous name and was not merged breaks, giving us a chance to review and check whether we really need to construct a cert pool or we can get away by using `netxlite`'s default pool (a static pool generated using `NewMozillaCertPool`).

Closes https://github.com/ooni/probe/issues/2444.

Co-authored-by: Simone Basso <bassosimone@gmail.com>