[backport] feat: lightweight throttling measurements (#1166)
This diff backports 7b886518d7929b454135674e0bb8dcc9388b621b from
the master branch to the release/3.18 branch.

This diff implements a lightweight approach to throttling that takes
advantage of the step-by-step design and should also be suitable to
measure throttling using `dslx` (see
https://github.com/ooni/probe/issues/2493).

Before discussing the approach implemented here, it is important to
point out that:

1. if we're using step-by-step, we're collecting up to 64 network events
for a single network connection;

2. with step-by-step, each trace is bound to a single network connection
or DNS round trip;

3. both Web Connectivity v0.5 and dslx use the step-by-step approach;

4. therefore, for extreme throttling of a single connection, 64 I/O
events are *a lot of events* to observe throttling;

5. additionally, we're currently limited at downloading `1<<19` bytes of
the body, so there is not much room for collecting *lots of data*
anyway;

6. additionally, if we were to collect more bytes, the bottleneck would
become collecting and uploading the HTTP response body to the OONI
backend.

That said, by exploiting the fact that step-by-step means that a trace
is bound to a single network connection, we can add passive atomic
collection of the bytes received by a trace. Because we're dealing with
unconnected UDP sockets, we also need to be careful about accounting the
bytes received from the peer that sent the bytes. To this end, we
maintain a map from the remote endpoint address and protocol to the
number of bytes received. The trace allows one to export the current
map. Because data collection is passive, we can start as late as the
HTTP download and we would still collect correct cumulative data.

We also introduce a new sampler for measuring throttling. The design of
the sampler is similar to the design we're using inside of ndt7. We use
a memoryless ticker to avoid sampling periodically but we clamp the
distribution such that we will typically receive the expected amount of
samplers for each time period.

It is also worth noting that I believe the already collected 64 network
events are fine to determine throttling, but we cannot know for sure,
hence it makes sense to improve our data collection capabilities.

The related spec PR is https://github.com/ooni/spec/pull/276.

Once this diff is merged, we would still need to do the following:

- [ ] update dslx to use this functionality
- [ ] land Web Connectivity LTE

The latter is fundamental to collect speed samples. We're not doing that
with Web Connectivity v0.4.

While there, this diff also improves the measurexlite documentation a
bit.