feat(netemx): redesign to simplify usage (#1168)
This commit shuffles around the `netemx` implementation to simplify
construction and usage. Here are some design goals that underpin this
set of changes:

1. construction using optional functions and sane defaults, so one needs
to write much less;

2. we always want an HTTP server listening 80/tcp, 443/tcp, and 443/udp,
so we can simplify the code *a lot*.
 
I worked on this change while trying to wrap up with
https://github.com/ooni/probe/issues/2493. My main aim was to add netem
based smoke testing to Web Connectivity LTE such that we can start
increasing the amount of users using it with some extra confidence
compared to the current situation where code coverage is very low.