refactor: move measurement-time functions to model (#1169)
This diff moves measurement-time functions to the internal/model
package. In turn, this work would be helpful to write some basic smoke
tests for Web Connectivity LTE. In turn, doing that is functional to
https://github.com/ooni/probe/issues/2493. With more smoke testing for
Web Connectivity LTE, we can enable it for more users and we can
therefore collect more data about potential throttling.