feat(netem): add custom TCP listener
We need this functionality to write tests for whatsapp and telegram, which
are using nonstandard services. Whatsapp requires a TCP listener speaking
a protocol that we don't know about (presumably XMPP with encryption). Telegram
listens for plaintext HTTP even when using port 443.

Part of https://github.com/ooni/probe/issues/2461.