[backport] feat(oonimkall): experimental OONI Run v2 API (#1176)
This commit backports 51b694e7ef04fe4770bb1b31aa25f5878ab38988.

This commit introduces an experimental OONI Run v2 API, which will
simplify the development of Android and iOS OONI Run v2 functionality.
We're using the testing host for now (i.e.,
https://ams-pg-test.ooni.org/).

This code is not suitable to be used in production. The API may change
without notice in the future.

The current API is the result of a trial and error process involving me
and @aanorbel.

Part of https://github.com/ooni/probe/issues/2503