refactor(oohelperd): move code out of package main (#1189)
This diff refactors the oohelperd implementation to move code out of
package main. Specifically, we move most of the code from
`./internal/cmd/oohelperd` to `./internal/oohelperd`. Thanks to this
refactoring, we can reuse the implementation of the oohelperd for
writing netem based integration tests.

Part of https://github.com/ooni/probe/issues/2461

---------

Co-authored-by: kelmenhorst <k.elmenhorst@mailbox.org>