feat(netxlite): decouple from tproxy singleton (#1196)
Code in netxlite is currently tightly coupled with the tproxy singleton
(i.e., with a `model.UnderlyingNetwork`). Unfortunately, to write netemx
tests for Web Connectivity v0.5, we need to break this coupling. In
fact, we need the test helper's code to use a different underlying
network that uses a different IP address and is not censored, unlike the
network used by the probe.

To this end, in this diff we're introducing a nil-safe provider for a
tproxy. When the provider struct pointer is nil or the tproxy field is
nil, the code behaves as before and uses the tproxy singleton.
Otherwise, the code will use the given tproxy. Then we modify
fundamental types to explicitly depend on this nil-safe provider rather
than implicitly depending on the tproxy singleton.

We're soon going to take advantage of this diff and introduce support
for constructing common netxlite types using a custom tproxy that is not
a singleton but rather is a field of a concrete "net" structure.

This work has been extracted and adapted from
https://github.com/ooni/probe-cli/pull/1185. The reference issue for
this pull request is https://github.com/ooni/probe/issues/2461.

---------

Co-authored-by: kelmenhorst <k.elmenhorst@mailbox.org>