refactor(netemx): factor function for listening
This diff factors the function used for listening out of the loop where we
create all the servers stacks. The same diff is part of the original PR that
introduced these changes: https://github.com/ooni/probe-cli/pull/1185.

The intent of pushing this commit is to make sure the diff we're merging
next is easier to understand because there's no refactoring noise.

Reference issue: https://github.com/ooni/probe/issues/2461.