fix(netemx): avoid suggesting more than one role is possible
This diff modifies the netemx scenario to avoid using a bit mask
since it's not possible for a host to have multiple roles.

While there, upgrade github.com/ooni/netem to pull a fix that causes
the start topology to return an error if you attempt to add the
same host's IP address multiple times.

Part of https://github.com/ooni/probe/issues/2461