feat(webconnectivityqa): add test for android_dns_cache_no_data (#1210)
Part of https://github.com/ooni/probe/issues/2029. The general idea is
to modify v0.4 in a subsequent PR to make it WAI for this test case.
Currently v0.4 does not correctly support this case, which is why we're
doing this work. (See also https://github.com/ooni/probe/issues/2499).