fix(webconnectivity): handle android_dns_cache_no_data as dns anomaly
This diff changes the handling of android_dns_cache_no_data to mark it and
any other DNS inconsistenct as DNS anomaly.

The previous implementation was not handling any DNS inconsistency as an
anomaly, therefore I needed to adjust a bunch of summary_test.go tests.

Part of https://github.com/ooni/probe/issues/2499 and
https://github.com/ooni/probe/issues/2029.

I bumped the version number because this is a significant change
in the analysis algorithm implemented by the probe.