chore(webconnectivityqa): import webconnectivity_tcpip_blocking_with_consistent_dns
This diff imports the given QA/webconnectivity.py integration test into the
webconnectivityqa framework and removes it from Python. To this end, we need
to modify how we handle timeouts in netxlite such that it is possible to
change the default timeout programmatically when using netemx.

Part of https://github.com/ooni/probe/issues/1803.