feat(netxlite): expose MaybeCustomUnderlyingNetwork
If we want to make other OONI components use either the default
model.UnderlyingNetwork or a custom one, we need to expose the
related routing functionality from netxlite.

Having this functionality handy would be useful, for example, to
write a resolver IP lookup package that would work as intended both
when using the regular network and when using netem.

In turn, this functionality opens up the possibility of running
the whole of ooniprobe inside of netem in an easier way.

Related to https://github.com/ooni/probe/issues/2461.