fix(netemx): limit using the 130.192.91.x address space
As we learned in https://github.com/ooni/probe-cli/pull/1216, using the
130.192.91.x namespace for every IP address in netemx breaks mapping domain
names to IP addresess in Web Connectivity. No IP address would ever, in
fact, be inconsistent, because they all belong to AS137.

Initially, I thought about overriding the code that maps IP addresses to
ASNs, to provide a custom implementation. But then I realized it was a
more thorough test to use the default implementation (relying on maxminddb
files) and using the correct IP addresses in the correct address space.

My original thought for using 130.192.91.x addresses was that they were
not the right addresses for the domains we're testing, thus, in the event
in which netem was not WAI, all tests would have failed. However, we
have many tests checking that netem is WAI already, so probably I was
being excessively paranoid.

As a result, this patch modifies the code to use the correct addresses.

Part of https://github.com/ooni/probe/issues/1803.