refactor(netemx): move notable IP addrs in a separate file
I am trying to lift some netemx restrictions that make certain Web Connectivity
integration tests a bit unrealistic. To progress towards this goal, I have
determined that I need to (1) improve the construction for the  QAEnv and (2)
make the scenario constructor a function that initialized a given QAEnv rather
than rolling additional abstraction on top of it.

The current diff moves IP addresses around to create space for myself to
attempt to implement the above changes.

The reference issue is https://github.com/ooni/probe/issues/1803.