feat(netemx): adapt NetStackServerFactory from telegram
The overall intent of this work is to unify how we manage HTTP, DNS, and
ordinary netstacks for netemx, and allow the same IP address to host multiple
servers rather than just one, as it's currently the case.

To this end, the current diff adapts code that used to be an integration
test for telegram to become the cleartext HTTP NetStackServerFactory.

Reference issue: https://github.com/ooni/probe/issues/1803