feat(netemx): HTTPS server implementing NetStackServerFactory
This diff continues improving and refactoring netemx with the objective
of unifying how we create all kind of servers.

Here, specifically, we modify the HTTP server implementing
NetStackServerFactory implemented in the previous commit and
obtain an HTTPS server honouring NetStackServerFactory.

Crucially, this diff also adds support for overriding the TLS config
passed to the server, which enables us to test for expired
certificates, self-signed certificates, and so forth.

Reference issue: https://github.com/ooni/probe/issues/1803.