feat(netemx): implement NetStackServerFactory for DNS-over-UDP
This diff moves forward the work we're doing to try to unify the way
in which all the possible servers in netemx are constructed.

The general idea is to enable configuring every kind of server
using the QAEnvOptionNetStack option.

To this end, we need to implement NetStackServerFactory.

In this diff, in particular, we're implementing a NetStackServerFactory
for constructing DNS-over-UDP servers. These servers would use:

1. the OtherResolversConfig for lookups (which makes sense because
the [MustNewQAEnv] internally constructs the ISP resolver);

2. the logger configured for the QAEnv.

Part of https://github.com/ooni/probe/issues/1803