fix(netemx): address issues with quic-go/quic-go
This diff fixes https://github.com/ooni/probe/issues/2527.

The comment I originally wrote in the code to explain the issue read:

> I am wondering whether I am not fully understanding how quic-go/quic-go works.

It turns out this comment was true. I was assuming http3.Server was closing
the listener, however, that was not the case.

I noticed this issue when reading again netemx/qaenv.go.

So, it all boiled down to PEBCAK 😬.