refactor(netemx): use QAEnvOptionNetStack for ScenarioRoleDNSOverHTTPS
This diff continues the refactoring of netemx to use a single mechanism
to create all the possible kind of servers.

While there, this diff removes the limitation that we cannot create
more than a single server per IP address, which was one of the reasons
why we started this refactoring process.

While there, make sure we have full coverage of netemx.

Part of https://github.com/ooni/probe/issues/1803