cleanup(netemx): QAEnvOptionHTTPServer is now syntactic sugar
With all the changes implemented so far, we're now well positioned
to futher cleanup netemx by implementing QAEnvOptionHTTPServer as
just syntactic sugar for QAEnvOptionNetStack.

This change allows us to delete more code that is duplicating
existing functionality and is not otherwise useful.

While there, introduce a role for a blockpage server and make sure
that such a server only listens for HTTP, thus fixing a case in which
the behavior of netemx was not consistent with the reality.

Part of https://github.com/ooni/probe/issues/1803.