feat: start merting filtering into testingx
To continue rewriting Jafar based tests to use netemx, the next
step is to adapt the proxies used by Jafar to be usable from within
netemx to implementing equivalent test cases.

In turn, netemx is using testingx for general-purpose test servers
that could also be interesting for other packages.

We also have the netxlite/filtering package, which is ~fine but
has a string-based API, where an interface-based API would be more
proper and easier to compose. (We historically use a string-based
API there because we previously attempted to replace Jafar with
code in userspace usinf netxlite/filtering.)

The first step in this quest is therefore to rewrite the DNS code
inside netxlite/filtering to belong to testingx.

Part of https://github.com/ooni/probe/issues/1803.