feat: start merging filtering into testingx (#1234)
## Checklist

- [x] I have read the [contribution
guidelines](https://github.com/ooni/probe-cli/blob/master/CONTRIBUTING.md)
- [x] reference issue for this pull request:
https://github.com/ooni/probe/issues/1803
- [x] if you changed anything related to how experiments work and you
need to reflect these changes in the ooni/spec repository, please link
to the related ooni/spec pull request: N/A
- [x] if you changed code inside an experiment, make sure you bump its
version number: N/A

## Description

To continue rewriting Jafar based tests to use netemx, the next step is
to adapt the proxies used by Jafar to be usable from within netemx to
implement equivalent test cases.

In turn, netemx is using testingx for general-purpose test servers that
could also be interesting for other packages.

We also have the netxlite/filtering package, which is ~fine but has a
string-based API, where an interface-based API would be more proper and
easier to compose. (We historically use a string-based API there because
we previously attempted to replace Jafar with code in userspace usinf
netxlite/filtering.)

The first step in this quest is therefore to rewrite the DNS code inside
netxlite/filtering and move it to testingx.

While there rename netemx.UDPResolverFactory to DNSOverUDPServerFactory
for consistency.