refactor: merge filtering's HTTP code with testingx (#1235)
## Checklist

- [x] I have read the [contribution
guidelines](https://github.com/ooni/probe-cli/blob/master/CONTRIBUTING.md)
- [x] reference issue for this pull request:
https://github.com/ooni/probe/issues/1803
- [x] if you changed anything related to how experiments work and you
need to reflect these changes in the ooni/spec repository, please link
to the related ooni/spec pull request: N/A
- [x] if you changed code inside an experiment, make sure you bump its
version number: N/A

## Description

This diff merges internal/filtering's code with testingx. While doing
this, we refactor the code to (1) be a drop-in replacement for the
net/http/httptest package and (2) use interfaces rather than string
values to select the proper behavior, which makes the code more
composable than it was before.

The overall objective of this effort is to consolidate filtering into
testingx such that we can later merge the proxy capabilities implemented
by Jafar into the testingx package, which is a stepping stone for
ditching Jafar.