refactor: finish merging filtering into testingx
This commit finishes the mini quest about merging netxlite/filtering
with testingx and replaces the string-based API with interfaces.

What remains to be done is now to implement specific handles that
mimic the proxies implemented by Jafar. Then, it remains to see how
to port the remaining QA tests to webconnectivityqa.

This work is part of https://github.com/ooni/probe/issues/1803.