webconnectivityqa: import test cases using proxies
This diff imports into webconnectivityqa test cases using proxies that
are currently also implemented by ./QA/webconnectivity.py.

The ./QA/webconnectivity.py contains three test cases. Two of them
deal with using transparent proxies without DNS lies, which was easy
to do using iptables, and much harder now. However, it's doubtful
whether those two cases are actually very useful, since there is no
measurement feature which allows us to distinguish them from what
we would otherwise get (perhaps, possibly latency?).

The third case, instead, is interesting and deals with the DNS serving
to users the IP addresses of transparent HTTP and TLS proxies. To
make this test case more similar to what it was in Python, and considering
that LTE uses many resolvers, here I have chosen to use DNS spoofing,
which may or may not be the best choice for LTE in the long term.

Yet, since the objective currently is to be able to check v0.4 against
webconnectivityqa and the A/B comparison and the focus to LTE will
come at a later stage, this seems good enough for now.

Part of https://github.com/ooni/probe/issues/1803.