refactor(measurexlite): move OperationLogger to logx
I am starting to look into how to support beacons into the codebase
as explained by https://github.com/ooni/probe/issues/2531.

The first issue that I stumbled upon is that there is excessive
coupling between packages at the lower-levels, which we can reduce
with easy refactoring, and which would help reasoning on what to
do to address the underlying issue.

This diff is the first minor step towards cleaning up the code
a bit: we are moving a functionality that pertains to logging
from measurexlite to logx (which is where it should be).

While there, I noticed I needed to run `go mod tidy` to update
the `go.mod` and `go.sum` files, which were outdated.