refactor(netxlite): use *Netx for the system resolver (#1248)
This diff modifies how we construct netxlite's system resolver such that
public functions use the *Netx equivalents.

While there, recognize that there wasn't enough testing for the optional
wrappers provided by `model.DNSTransportWrapper` and that we are not
using this functionality. So, rather than writing new tests for this
functionality, we can actually just drop it and simplify the codebase.

While there, recognize that `netxlite.WrapDNSTransport` could easily
become private.

While there, recognize that `./legacy/netx` needs lots of public
function being exported by `netxlite` but we don't need to expose that
many implementation details to new code that will be using `*Netx`. So,
make sure all the new methods we create for `*Netx` are actually private
methods. (Ideally, the API surface of `netxlite` should be smaller; we
would not be able to get there for quite some time, but we can at least
avoid increasing the API surface.)

The general idea for which I am pushing here is to have additional
clarity about dependencies, to better analyze the requirements of non
measuring code for https://github.com/ooni/probe/issues/2531.