refactor(netxlite): use *Netx for creating QUIC dialers
This diff is like https://github.com/ooni/probe-cli/commit/8a5edc22e3fde71e412715e6f96797376252b0ce
but here we use *Netx to create QUIC dialers.

The general idea of this patchset is to ensure we're not
using duplicate code for constructing netxlite types, which
is good to do now, because we're about to introduce new
netxlite types for the network with which we communicate
with the OONI backend.

Reference issue: https://github.com/ooni/probe/issues/2531