refactor(netxlite): use *Netx for creating HTTP3 transports
This diff is like
https://github.com/ooni/probe-cli/commit/50279a7f659d407c2dab6904227083087ae745e7
but uses *Netx to create HTTP3 transports.

The general idea of this patchset is to ensure we're not using duplicate
code for constructing netxlite types, which is good to do now, because
we're about to introduce new netxlite types for the network with which
we communicate with the OONI backend.

Reference issue: https://github.com/ooni/probe/issues/2531