feat(netxlite): use *Netx for creating DoH resolvers (#1255)
While working on removing technical debt, as part of
https://github.com/ooni/probe/issues/2531, it makes sense ensuring that
there is a single way of constructing netxlite types by always using an
instance of *Netx. This job is relatively easy because we are a couple
of patches away from achieving this goal and, by doing this, we would
avoid creating future technical debt. So, I am going to ahead and
implement this.

Also, apparently, the oohelperd uses this functionality and
somehow we missed doing this when converting it to using netemx. We
should fix this soon. (This issue immediately validates what I was
saying above about techdebt 😅)